+++ app/manage/[id]/Manage.module.css
@@ -125,6 +125,50 @@
   }
 }
 
+/* toast 관련 스타일 */
+.toast-container {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+  align-items: center;
+}
+
+.toast-message {
+  font-weight: 500;
+}
+
+.toast-btn-group {
+  display: flex;
+  gap: 8px;
+}
+
+.toast-btn-confirm {
+  padding: 6px 16px;
+  background-color: var(--bg-navy);
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+}
+
+.toast-btn-reject {
+  padding: 6px 16px;
+  background-color: var(--bg-red);
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+}
+
+.toast-btn-cancel {
+  padding: 6px 16px;
+  background-color: #e5e7eb;
+  color: #374151;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+}
+
 /* 푸터를 감싸는 div */
 .footer-wrapper {
   display: none;
